test: add hybrid package integration test
